Split code related to API and DB calls in data_sources.py and models.py respectively.
